pyproject.toml
[project]
name = "nano-banana-hack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-genai>=1.33.0",
    "gradio>=3.36.1",
    "ipykernel",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.18",
    "mkdocstrings>=0.30.0",
    "mkdocstrings-python>=1.18.2",
]
